mesh: deep changes avoiding bug case 3
In short: a root vertex is non-placed in space, but is linked to each of
the corner. Thus, there is four faces and eight edges in the outside
objects, however they remain hidden for client code.

This trick avoid any vertex to not be linked to at least three other
vertices. Also, as the four OutsideFace instances does not collide
anything, it's a elegant dead-end for any delaunay application search.